requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-Login>=0.6.3
Flask-Migrate>=4.0.5
Flask-SocketIO>=5.3.5
Werkzeug>=3.0.1
SQLAlchemy>=2.0.35
stripe>=7.8.0
python-dotenv>=1.0.0
Pillow>=10.4.0
email-validator>=2.1.0
gunicorn>=22.0.0
eventlet>=0.36.1
